Move common validation interfaces and utilities in proper place
