+++ src/forms/createPortalSchema.ts
@@ -1,6 +1,7 @@
 import * as Yup from "yup";
 
-import { PortalSchema } from "forms/common";
+import { PortalSchema } from "types/validation";
+
 import { roomUrlSchema } from "forms/roomUrlSchema";
 import { venueNameSchema } from "forms/venueNameSchema";
