+++ src/forms/createSpaceSchema.ts
@@ -1,6 +1,7 @@
 import * as Yup from "yup";
 
-import { SpaceSchema } from "forms/common";
+import { SpaceSchema } from "types/validation";
+
 import { venueNameSchema } from "forms/venueNameSchema";
 
 export const createSpaceSchema = Yup.object().shape<SpaceSchema>({
