+++ src/types/validation.ts
@@ -1,0 +1,8 @@
+export interface SpaceSchema {
+  template?: string;
+  venueName?: string;
+}
+
+export interface PortalSchema extends SpaceSchema {
+  roomUrl?: string;
+}
