+++ src/utils/getLinkFromText.tsx
@@ -2,6 +2,7 @@ import React from "react";
 
 import { getExtraLinkProps, isExternalUrl } from "./url";
 
+// @debt check if used anywhere and remove if not
 export const getLinkFromText = (text: string) => {
   const urlRegex = /[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_+.~#?&/=]*)/g;
   return text.split(" ").map((word, index) => {
