+++ src/utils/validation.ts
@@ -2,15 +2,6 @@ import firebase from "firebase/app";
 
 import { createSlug } from "api/admin";
 
-export interface SpaceSchema {
-  template?: string;
-  venueName?: string;
-}
-
-export interface PortalSchema extends SpaceSchema {
-  roomUrl?: string;
-}
-
 export const messageMustBeMinimum = (fieldName: string, min: number) =>
   `${fieldName} must be at least ${min} characters`;
